cloud-functions/get_signed_url/requirements.txt
google-cloud-storage
Flask
firebase-admin
six
google-auth